Fix haproxy not having its certs properly

+++ roles/haproxy/templates/cert_update.sh
@@ -3,7 +3,9 @@
 {{ ansible_managed|comment }}
 
 for path in /etc/letsencrypt/live/*; do
-  cat ${path}/{fullchain,privkey}.pem > "/etc/haproxy/certs/$(basename "$path").pem"
+  if [[ -f ${path}/fullchain.pem ]]; then
+    cat ${path}/{fullchain,privkey}.pem > "/etc/haproxy/certs/$(basename "$path").pem"
+  fi
 done
 
 systemctl is-active --quiet haproxy.service && systemctl reload haproxy.service
